refactor tests and stories to reflect renaming
